requirements.txt
opencv-python
ultralytics
numpy
pandas
matplotlib
scikit-image
sort-tracker